% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{all}
\alias{all}
\title{Fit a logistic regression model with lasso using glmnet() function}
\usage{
all(X, y)
}
\arguments{
\item{X}{input matrix.}

\item{y}{response variable.}
}
\description{
Fit a logistic regression model with lasso using glmnet() function. Returns
the computing time and misclassification rate in a vector form.
}
