% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_generate.R
\name{beta_generate}
\alias{beta_generate}
\title{Create a sparse beta}
\usage{
beta_generate(p, k, min, max)
}
\arguments{
\item{p}{the desired number of columns.}

\item{k}{the desired number of non-zero entries.}

\item{min}{lower limit of the distribution.}

\item{max}{upper limit of the distribution.}
}
\description{
Create a sparse beta in matrix form with non-zero entries randomly drawn from
uniform distribution.
}
\examples{
beta_generate(p=1000,k=30,min=.2,max=.8)

}
