% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_generate.R
\name{y_generate}
\alias{y_generate}
\title{Create a response variable y}
\usage{
y_generate(beta, X)
}
\arguments{
\item{X}{input matrix, each row is an observation.}

\item{\eqn{\beta}}{coefficient vector.}
}
\description{
Create a response variable y follows Bernoulli distribution with success probability
\eqn{p(X\beta)} given \eqn{X} and \eqn{\beta}.
}
\examples{
n<-100
p<-200
mu<-matrix(rep(0,p),nrow=p)
sigma<-diag(1,nrow=p)
X<-mvrnorm(n,mu,sigma)
beta<-beta_generate(p,k=30,min=.2,max=.8)
y_generate(beta,X)

}
